% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_euphoria.R
\docType{package}
\name{euphoria-package}
\alias{euphoria}
\alias{euphoria-package}
\title{euphoria: ggplot2 Themes and Color Palette Inspired by HBO's Euphoria}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a ggplot2 theme and color palette inspired by the visual aesthetic of HBO's television show Euphoria. Features deep purple-black backgrounds, neon orchid accents, and character-named color palettes for use in data visualizations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/samracuna/euphoria}
  \item Report bugs at \url{https://github.com/samracuna/euphoria/issues}
}

}
\author{
\strong{Maintainer}: Sam Acuna \email{samruthacuna@gmail.com}

}
\keyword{internal}
