% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_euphoria.R
\docType{data}
\name{euphoria_palette_discrete}
\alias{euphoria_palette_discrete}
\title{Discrete Euphoria palette (8 colors for categorical data)}
\format{
An object of class \code{character} of length 8.
}
\usage{
euphoria_palette_discrete
}
\description{
Discrete Euphoria palette (8 colors for categorical data)
}
\keyword{datasets}
