% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_euphoria.R
\docType{data}
\name{euphoria_palette_seq}
\alias{euphoria_palette_seq}
\title{Sequential Euphoria palette (void to glitter, for continuous data)}
\format{
An object of class \code{character} of length 7.
}
\usage{
euphoria_palette_seq
}
\description{
Sequential Euphoria palette (void to glitter, for continuous data)
}
\keyword{datasets}
