% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_euphoria.R
\name{scale_colour_euphoria}
\alias{scale_colour_euphoria}
\alias{scale_color_euphoria}
\title{Euphoria discrete colour scale}
\usage{
scale_colour_euphoria(...)

scale_color_euphoria(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{ggplot2::scale_colour_manual}.}
}
\description{
Euphoria discrete colour scale
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg, colour = factor(cyl))) +
  geom_point() +
  scale_colour_euphoria()
}
