% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_euphoria.R
\name{scale_colour_euphoria_c}
\alias{scale_colour_euphoria_c}
\alias{scale_color_euphoria_c}
\title{Euphoria continuous colour scale}
\usage{
scale_colour_euphoria_c(...)

scale_color_euphoria_c(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{ggplot2::scale_colour_gradientn}.}
}
\description{
Euphoria continuous colour scale
}
