% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_euphoria.R
\name{scale_fill_euphoria}
\alias{scale_fill_euphoria}
\title{Euphoria discrete fill scale}
\usage{
scale_fill_euphoria(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{ggplot2::scale_fill_manual}.}
}
\description{
Euphoria discrete fill scale
}
\examples{
library(ggplot2)
ggplot(mpg, aes(class, fill = class)) +
  geom_bar() +
  scale_fill_euphoria()
}
