% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_euphoria.R
\name{scale_fill_euphoria_c}
\alias{scale_fill_euphoria_c}
\title{Euphoria continuous fill scale}
\usage{
scale_fill_euphoria_c(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{ggplot2::scale_fill_gradientn}.}
}
\description{
Euphoria continuous fill scale
}
\examples{
library(ggplot2)
ggplot(faithfuld, aes(waiting, eruptions, fill = density)) +
  geom_tile() +
  scale_fill_euphoria_c()
}
